+++ js/abstract-view.js
@@ -1,0 +1,29 @@
+export default class AbstractView {
+  constructor() {
+    if (new.target === AbstractView) {
+      throw new Error(`Can't instantiate AbstractView, only concrete one`);
+    }
+  }
+
+  get template() {
+    throw new Error(`Template is required`);
+  }
+
+  render() {
+    const wrapper = document.createElement(`div`);
+    wrapper.innerHTML = this.template.trim();
+    return wrapper.firstChild;
+  }
+
+  bind() {
+  }
+
+  get element() {
+    if (this._element) {
+      return this._element;
+    }
+    this._element = this.render();
+    this.bind(this._element);
+    return this._element;
+  }
+}
