+++ js/screens/result-controller.js
@@ -1,0 +1,5 @@
+import ResultView from '../views/result-view';
+
+const resultScreen = new ResultView();
+
+export default resultScreen;
