+++ js/screens/welcome-controller.js
@@ -1,0 +1,6 @@
+import WelcomeView from '../views/welcome-view';
+
+const welcomeScreen = new WelcomeView();
+welcomeScreen.onStartClick = () => window.showFirstGameScreen();
+
+export default welcomeScreen;
