+++ js/views/result-view.js
@@ -1,0 +1,27 @@
+import AbstractView from '../abstract-view';
+import {declension} from '../utils';
+import {resultsText} from '../game';
+
+export default class ResultView extends AbstractView {
+  constructor() {
+    super();
+  }
+
+  get template() {
+    const state = window.gameState;
+
+    return `<section class="main main--result">
+      <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
+
+      <h2 class="title">Вы настоящий меломан!</h2>
+      <div class="main-stat">За&nbsp;3&nbsp;минуты и 25&nbsp;секунд
+        <br>вы&nbsp;набрали ${state.totalPoints} балл${declension(state.totalPoints, ``, `а`, `ов`)} (8 быстрых)
+        <br>совершив ${state.wrongAnswers} ошиб${declension(state.wrongAnswers, `ку`, `ки`, `ок`)}</div>
+      <span class="main-comparison">
+        ${resultsText(state.otherPlayersResults, window.gameState)}
+      </span>
+      <span role="button" tabindex="0" class="main-replay">Сыграть ещё раз</span>
+    </section>`;
+  }
+
+}
